Chore: Renamed `src/cli` to `src/utils`
To leave that module name to what will be an actual CLI
